make fields required
When a field is not specified as required, by default the generated schema treats it as optional. In such case, when converting the schema back to struct, tools like go-jsonschema will make the field as a pointer type. Hence, a plain string field becomes *string, which is in-consistent. Adding the `required` tag fix the problem.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>